Refactor redux to work at top level
This connects the redux JSON fetching method to the LabelWrapper component, and allows it to take in a filename as an argument. This way, we can add blobs to the `static/` folder, and dynamically request them/create new label pages by passing the filename as a prop to LabelWrapper from the page.